Component renamed. Fixes a name conflict with the xiaomi aqara gateway component.
