+++ xiaomiplug/__init__.py
@@ -1,4 +1,4 @@
 # flake8: noqa
 from .protocol import Message, Utils
 from .containers import VacuumStatus, ConsumableStatus, CleaningDetails, CleaningSummary, Timer, AirQuality
-from .xiaomi import Device
+from .device import Device
