+++ xiaomiplug/version.py
